[Man] Amend error in the FetchUsers(3go) manual
The manual stated that in the case of an error the `FetchUser()`
function will return a tuple containig `nil` and an error. In actuality
the function will return a blank `User` struct and an error. This has
been corrected in the manual. Additionally the section on the return
value for `FetchUsers()` has been updated to be more in line with what
is said for `FetchUser()`.