[Users] Prefer ints over uints for max and counts
Technically, unsigned integers are prefered because there is no real
reason for these numbers to be negative. That being said there is no
real scenario where the range of these numbers should matter and using
signed integers instead of unsigned integers allows for much easier to
look at code as you can get rid of all the type casts to and from uint
and uint64 and all that junk.