[Users] Support personal bests
The `User.PersonalBests()` method allows you to grab all of a users
personal bests. If you want to filter the PBs by game, series, or
placement you can do so via the provided `UserFilter` struct. The struct
is optional though and can be nil. Additionally the method supports
embedding via the `UserFilter`.