[README] Link to files when mentioned
Previously when mentioning the `man/` directory and the projects
license, the mentions of these local files and directories were just
highlighted through either the use of inline code blocks or boldened
text. Instead we can make them link to the actual files and directories
they correspond to so the user can just click the links.