[Meta] Set vi textwidth to 100 in documentation
Vi/Vim/Whatever allows you to set a text width to make text
automatically wrap after a certain length. This is useful for
documentation files such as the README and manual pages as it can be
tedious to track how long each line is and to keep things nicely
formatted.