[Users] Add WorldRecords and Podiums as methods
The `User.WorldRecords()` method will offer a clean way to easily
retrieve all of a users world record runs. This method does not accept a
filter unlike `User.PersonalBests()` which should still be used if you
want to get WRs based off of a filter. The point of this method is
simply to make code that fetches WRs look cleaner. Likewise, the
`User.Podiums()` method retrieves all of a users top-3 times, also
without a PBFilter as an argument and for the same reasons.