[Users] Make FetchUser() take a string parameter
Previously, `FetchUser()` was a 1:1 copy of `FetchUsers()` except it
returned one user instead of a slice of users. This commit makes it so
that it now takes a string argument (a name) which it looks up instead
of taking a UserFilter struct. This is because the goal of the
`FetchUser()` function is to make code simpler, and the vast majority of
the time that it gets called will be with the goal of looking up a user
based on username.

Along with the code changes, the documentation for the function has also
been updated and the examples in other manuals that call the
`FetchUser()` function have also been updated to take the username as a
direct parameter.