[Http] Support request headers and return errors
This commit both adds support for specifying request headers, and
returns errors properly when speedrun.com returns a status code >= 400
with the exception of 420 (the rate limit).

Request headers can now be passed as a 3rd argument to
`requestAndUnmarshall()` or as a 2nd argument to `request()`. The map
key is the header name and the value is the header value (obviously).

Additionally, if the site responds with a status code >= 400 that is not
420 it now unmarshals the resulting JSON so that we can get nicer errors
returned. Finally after being rate limited 10 times the `request()`
function just returns with an error. This has to be done because some
absolute retard decided to make 420 be both a rate limit AND an error
that can occur when the site is under pressure.