[Make] Delete temporary files on exit
If the user attempts to run `make docs` without the correct permissions
the command will fail and the `tmp.tmac` and `tmp.3go` files will be
left in the working directory. By setting a trap for `EXIT` it can be
assured that in the case of a failure these temporary files will be
cleaned up.